setting : 레디스 설정
