--- src/main/java/kr/hhplus/be/server/config/RedisCacheConfig.java
@@ -1,34 +1,0 @@
-package kr.hhplus.be.server.config;
-
-import org.springframework.cache.annotation.EnableCaching;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.cache.RedisCacheConfiguration;
-import org.springframework.data.redis.cache.RedisCacheManager;
-import org.springframework.data.redis.connection.RedisConnectionFactory;
-
-import java.time.Duration;
-import java.util.HashMap;
-import java.util.Map;
-
-@Configuration
-@EnableCaching
-public class RedisCacheConfig {
-
-    @Bean
-    public RedisCacheManager cacheManager(RedisConnectionFactory redisConnectionFactory) {
-        RedisCacheConfiguration defaultConfig = RedisCacheConfiguration.defaultCacheConfig()
-            .entryTtl(Duration.ofMinutes(5));
-
-        Map<String, RedisCacheConfiguration> cacheConfigurations = new HashMap<>();
-
-        // 인기 상품 Product Id는 1일 유지
-        cacheConfigurations.put("TopSellingProductIds", RedisCacheConfiguration.defaultCacheConfig()
-            .entryTtl(Duration.ofHours(25)));
-
-        return RedisCacheManager.builder(redisConnectionFactory)
-            .cacheDefaults(defaultConfig)
-            .withInitialCacheConfigurations(cacheConfigurations)
-            .build();
-    }
-}
